Finish rename BackwardCompatiblityTest => OldFileVersionCompatibilityTests
