+++ src/OldFileVersionCompatibilityTests/OldFileVersionCompatibilityTests.csproj
@@ -9,8 +9,8 @@
     <ProjectGuid>{21F31760-D98D-41E6-BB7A-52619ED460C5}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Axe.Windows.BackwardCompatiblityTests</RootNamespace>
-    <AssemblyName>Axe.Windows.BackwardCompatiblityTests</AssemblyName>
+    <RootNamespace>Axe.Windows.OldFileVersionCompatibilityTests</RootNamespace>
+    <AssemblyName>Axe.Windows.OldFileVersionCompatibilityTests</AssemblyName>
     <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
